Viewer now shows all events, highlighting a selected event, showing the state of the Concert aggregate as of the selected event

+++ src/test/java/dev/ted/jitterticket/eventsourced/application/ConcertSummaryProjectorTest.java
@@ -100,7 +100,7 @@ class ConcertSummaryProjectorTest {
         ConcertRescheduled concertRescheduled = new ConcertRescheduled(concertId, 2, originalShowDateTime.plusMonths(2).plusHours(1), originalDoorsTime.plusHours(1));
         concertStore.save(concertId, List.of(concertScheduled, ticketsSold, concertRescheduled));
 
-        long desiredEventSequenceNumber = ticketsSold.eventSequence();
+        int desiredEventSequenceNumber = ticketsSold.eventSequence();
         var concertWithEvents = concertSummaryProjector.concertWithEventsThrough(concertId, desiredEventSequenceNumber);
 
         assertThat(concertWithEvents.concertEvents())
